Notifications proxy of trip, position and user

+++ notifications/src/main/java/be/vinci/ipl/notifications/NotificationsService.java
@@ -1,14 +1,23 @@
 package be.vinci.ipl.notifications;
 
+import be.vinci.ipl.notifications.data.NotificationsRepository;
+import be.vinci.ipl.notifications.data.TripsProxy;
+import be.vinci.ipl.notifications.data.UsersProxy;
+import be.vinci.ipl.notifications.models.Notification;
+import be.vinci.ipl.notifications.models.User;
 import org.springframework.stereotype.Service;
 
 @Service
 public class NotificationsService {
 
     private final NotificationsRepository repository;
+    private final UsersProxy usersProxy;
+    private final TripsProxy tripsProxy;
 
-    public NotificationsService(NotificationsRepository repository){
+    public NotificationsService(NotificationsRepository repository, UsersProxy usersProxy, TripsProxy tripsProxy){
         this.repository = repository;
+        this.usersProxy = usersProxy;
+        this.tripsProxy = tripsProxy;
     }
 
     /**
@@ -24,34 +33,44 @@ public class NotificationsService {
     /**
      * Reads all notifications from a specific user in repository
      * @param user the user to search for
-     * @return all notifications from this user
+     * @return all notifications from this user or false if user doesn't exist
      */
     public Iterable<Notification> readFromUser(long user){
+        if (usersProxy.readOne(user) == null) return null;
         return repository.findByUser(user);
     }
 
     /**
      * Deletes all notifications from a specific user from repository
      * @param user the user to delete the notifications from
+     * @return True if succesfuly deleted, false if user doesn't exist
      */
-    public void deleteFromUser(long user) {
+    public boolean deleteFromUser(long user) {
+        if (usersProxy.readOne(user) == null) return false;
         repository.deleteByUser(user);
+        return true;
     }
 
     /**
      * Reads all notifications from a specific trip in repository
      * @param trip the trip to search for
-     * @return all notifications frm this trip
+     * @return all notifications frm this trip or null if trip doesn't exist
      */
     public Iterable<Notification> readFromTrip(long trip) {
+        int tripInt = Math.toIntExact(trip);
+        if (tripsProxy.readOne(tripInt) == null) return null;
         return repository.findByTrip(trip);
     }
 
     /**
      * Deletes all notifications from a specific trip from repository
      * @param trip the trip to delete the notifications from
+     * @return True if succesfuly deleted, false if trip doesn't exist
      */
-    public void deleteFromTrip(long trip) {
+    public boolean deleteFromTrip(long trip) {
+        int tripInt = Math.toIntExact(trip);
+        if (tripsProxy.readOne(tripInt) == null) return false;
         repository.deleteByTrip(trip);
+        return true;
     }
 }
